9. Автоматизация для Яндекс.Дзен/requirements.txt.txt
Необходимые для работы библиотеки.

pandas, 
dash, 
dash_core_components, 
dash_html_components, 
sys, 
getopt,
plotly.graph_objs,
 
Из dash.dependencies операторы Input, Output
Из sqlalchemy - create_engine
Из datetime - datetime 


